requirements.txt
Flask==2.2.5
flask-cors
ujson==5.6.0
flasgger
parse
mongoengine
flask-mongoengine
pytz
# Flask
Flask-Swagger
PyJWT
# awslambdaric
# awsgi
# mangum
# flask-lambda
# serverless-wsgi>=2.0.2
requests>=2.25.1